requirements.txt
fastapi
uvicorn
google-genai
google-cloud-bigquery
google-auth
python-dotenv